refactor(ValidationAnimationContainer): moved to internal
