+++ src/deprecated/ChoiceFieldset/ChoiceFieldset.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import {Box, useSSRSafeId} from '../..'
 import createSlots from '../utils/create-slots'
 import {FormValidationStatus} from '../../utils/types/FormValidationStatus'
-import ValidationAnimationContainer from '../../_ValidationAnimationContainer'
+import ValidationAnimationContainer from '../../internal/components/ValidationAnimationContainer'
 import InputValidation from '../../_InputValidation'
 import ChoiceFieldsetListItem from './ChoiceFieldsetListItem'
 import ChoiceFieldsetDescription from './ChoiceFieldsetDescription'
