+++ src/internal/components/CheckboxOrRadioGroup/CheckboxOrRadioGroup.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import Box from '../../../Box'
-import ValidationAnimationContainer from '../../../_ValidationAnimationContainer'
+import ValidationAnimationContainer from '../ValidationAnimationContainer'
 import {get} from '../../../constants'
 import {useSSRSafeId} from '../../../utils/ssr'
 import CheckboxOrRadioGroupCaption from './CheckboxOrRadioGroupCaption'
